feat: refactor lambda function for nexmark
