+++ bench/Cargo.toml
@@ -15,6 +15,8 @@ datafusion = { git = "https://github.com/flock-lab/arrow-datafusion", branch = "
 driver = { path = "../src/driver" }
 env_logger = "^0.9"
 futures = "0.3"
+humantime = "2.1.0"
+
 itertools = "0.10.0"
 lazy_static = "1.4"
 log = "0.4.14"
@@ -26,7 +28,6 @@ rusoto_s3 = "0.47.0"
 serde_json = "1.0"
 structopt = { version = "0.3", default-features = false }
 tokio = { version = "1.4", features = [ "macros", "io-util", "sync", "rt-multi-thread", "time" ] }
-humantime = "2.1.0"
 
 [[bin]]
 name = "nexmark_bench"
