+++ src/driver/Cargo.toml
@@ -18,15 +18,15 @@ blake2 = "0.9"
 chrono = "0.4.19"
 daggy = { git = "https://github.com/flock-lab/daggy", branch = "master" }
 datafusion = { git = "https://github.com/flock-lab/arrow-datafusion", branch = "flock" }
+env_logger = "^0.9"
 lazy_static = "1.4"
+log = "0.4.14"
 runtime = { path = "../../src/runtime" }
 rusoto_core = "0.47.0"
 rusoto_iam = "0.47.0"
 rusoto_kinesis = "0.47.0"
 rusoto_lambda = "0.47.0"
 rusoto_logs = "0.47.0"
-log = "0.4.14"
-env_logger = "^0.9"
 tokio = { version = "1.4", features = [ "macros", "io-util", "sync", "rt-multi-thread" ] }
 
 # A list of all of the optional dependencies, some of which are included in the
